+++ src/Filesystems/S3.php
@@ -6,26 +6,35 @@ use Aws\CloudFront\CloudFrontClient;
 use Aws\CloudFront\Exception\CloudFrontException;
 use Aws\Credentials\CredentialProvider;
 use Aws\Credentials\Credentials;
+use Aws\Exception\CredentialsException;
 use Aws\Rekognition\RekognitionClient;
+use Aws\S3\S3Client;
 use Aws\Sts\StsClient;
-use CraftCms\AwsS3\Assets\AwsS3Bundle;
-use CraftCms\AwsS3\Enums\BucketSelectionMode;
 use CraftCms\AwsS3\Events\InvalidatingPaths;
 use CraftCms\AwsS3\Exceptions\FaceDetectionException;
 use CraftCms\Cms\Asset\Elements\Asset;
+use CraftCms\Cms\Cp\SelectOptions;
 use CraftCms\Cms\Filesystem\Filesystems\Filesystem;
-use CraftCms\Cms\Shared\Enums\TimePeriod;
+use CraftCms\Cms\Form\Controls\Combobox;
+use CraftCms\Cms\Form\Controls\Lightswitch;
+use CraftCms\Cms\Form\Controls\Text;
+use CraftCms\Cms\Form\Form;
+use CraftCms\Cms\Form\FormContext;
+use CraftCms\Cms\Form\Nodes\Callout;
+use CraftCms\Cms\Form\Nodes\Field;
+use CraftCms\Cms\Form\Nodes\Heading;
+use CraftCms\Cms\Form\Nodes\Separator;
 use CraftCms\Cms\Support\Arr;
 use CraftCms\Cms\Support\Env;
 use CraftCms\Cms\Support\Str;
-use CraftCms\Cms\View\LegacyAssets\InternalAssetRegistry;
+use CraftCms\Cms\Validation\Rules\EnvValueRule;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Crypt;
 use Illuminate\Support\Facades\Log;
 use League\Flysystem\Visibility;
 use Override;
 use function CraftCms\Cms\t;
-use function CraftCms\Cms\template;
 
 class S3 extends Filesystem
 {
@@ -34,6 +43,7 @@ class S3 extends Filesystem
     const string AWS_DEFAULT_REGION = 'us-east-1';
     const int AWS_REKOGNITION_MIN_CONFIDENCE = 80;
 
+    public ?string $url = null;
     public ?string $keyId = null;
     public ?string $secret = null;
     public string $region = self::AWS_DEFAULT_REGION;
@@ -41,7 +51,6 @@ class S3 extends Filesystem
     public ?string $expires = null;
     public ?string $subfolder = null;
     public bool $makeUploadsPublic = true;
-    public BucketSelectionMode $bucketSelectionMode = BucketSelectionMode::Choose;
 
     public ?string $cfDistributionId = null;
     public ?string $cfPrefix = null;
@@ -53,14 +62,18 @@ class S3 extends Filesystem
 
     public function __construct(object|array $config = [])
     {
+        // The plugin may be initialized with legacy configuration values!
+        // Bucket and region are now both “manually” set, with the help of suggested values;
+        // we don’t need to track `manualBucket` or `manualRegion` separately, and `bucketSelectionMode`
+        // was only used to preserve UI state in the plugin’s settings screen.
         if ($config['manualBucket'] ?? false) {
-            if (isset($config['bucketSelectionMode']) && $config['bucketSelectionMode'] === BucketSelectionMode::Manual) {
+            if (isset($config['bucketSelectionMode']) && $config['bucketSelectionMode'] === 'manual') {
                 $config['bucket'] = Arr::pull($config, 'manualBucket');
                 $config['region'] = Arr::pull($config, 'manualRegion');
             }
         }
 
-        unset($config['manualBucket'], $config['manualRegion']);
+        unset($config['manualBucket'], $config['manualRegion'], $config['bucketSelectionMode']);
 
         parent::__construct($config);
     }
@@ -86,7 +99,7 @@ class S3 extends Filesystem
                 $expires = $now->add($duration);
 
                 $options['CacheControl'] = sprintf('max-age=%d', $expires->getTimestamp() - $now->getTimestamp());
-            } catch (\InvalidArgumentException $e) {
+            } catch (\DateMalformedIntervalStringException $e) {
                 Log::warning(sprintf('Skipped setting `options.CacheControl` when configuring an AWS S3 disk due to an invalid duration string: %s', $this->expires));
             }
         }
@@ -94,9 +107,8 @@ class S3 extends Filesystem
         return $this->getClientConfig() + [
             'driver' => 's3',
             'bucket' => $this->bucket,
-            'url' => Env::parse($this->url),
-            // The `prefix` option is eventually passed to the S3 adapter as `root`:
-            'prefix' => Env::parse($this->subfolder),
+            'url' => $this->getRootUrl(),
+            'root' => Env::parse($this->subfolder),
             'visibility' => $this->makeUploadsPublic ? Visibility::PUBLIC : Visibility::PRIVATE,
             'use_path_style_endpoint' => env('AWS_USE_PATH_STYLE_ENDPOINT', false),
             'throw' => false,
@@ -107,27 +119,149 @@ class S3 extends Filesystem
 
     public function getRules(): array
     {
-        return parent::getRules() + [];
+        return parent::getRules() + [
+            'url' => [
+                new EnvValueRule(['url']),
+            ],
+            'expires' => [
+                'string',
+                function (string $attribute, mixed $value, \Closure $fail) {
+                    try {
+                        // All we need to do is make sure it compiles:
+                        \DateInterval::createFromDateString($value);
+                    } catch (\DateMalformedIntervalStringException $e) {
+                        $fail(t('The cache duration must be a valid date interval string.'));
+                    }
+                },
+            ],
+        ];
     }
 
-    public function getSettingsHtml(): ?string
+    public function settingsForm(FormContext $context = new FormContext): ?Form
     {
-        app(InternalAssetRegistry::class)->register(AwsS3Bundle::class);
+        $environmentTip = sprintf(
+            '%s [%s](%s)',
+            t('Type `$` to choose an environment variable.'),
+            t('Learn more'),
+            'https://craftcms.com/docs/5.x/configure.html#control-panel-settings',
+        );
+        $expansions = SelectOptions::getEnvTextExpanderTriggers();
 
-        return template('aws-s3/fsSettings', [
-            'fs' => $this,
-            'periods' => array_merge(
-                ['' => ''],
-                Arr::mapWithKeys(TimePeriod::cases(), fn ($p) => [$p->value => $p->label()]),
-            ),
-            'bucketSelectionModes' => Arr::mapWithKeys(BucketSelectionMode::cases(), fn ($mode) => [$mode->value => $mode->label()]),
-        ]);
+        $settingsForm = Form::make();
+
+        $settingsForm->add(Field::make(t('Base URL'))
+            ->instructions(t('The base URL to the files in this filesystem. See the AWS documentation on [website endpoints]({url}) for more information. Leave blank if you don’t want Craft to generate URLs for assets on this filesystem.', ['url' => 'https://docs.aws.amazon.com/AmazonS3/latest/userguide/WebsiteEndpoints.html'], 'aws-s3'))
+            ->control(Text::make('url')
+                ->textExpanderTriggers(SelectOptions::getEnvTextExpanderTriggers(true, fn ($value): bool => Str::isUrl($value)))
+                ->placeholder('https://s3.your-region.amazonaws.com/your-bucket-name/'))
+            ->tip(t('Type `$` to choose an environment variable, or `@` to choose an alias.')));
+
+        $settingsForm->add(Separator::make('credentials-separator'));
+
+        $settingsForm->add(
+            Field::make(t('Access Key ID', category: 'aws-s3'), Text::make('keyId')
+                ->textExpanderTriggers($expansions))
+                ->instructions(t('You can leave this field empty if you are using an EC2 instance with an applicable IAM role assignment.', category: 'aws-s3'))
+                ->tip($environmentTip),
+            Field::make(t('Secret Access Key', category: 'aws-s3'), Text::make('secret')
+                ->textExpanderTriggers($expansions))
+                ->instructions(t('You can leave this field empty if you are using an EC2 instance with an applicable IAM role assignment.', category: 'aws-s3'))
+                ->tip($environmentTip),
+            Field::make(t('Region', category: 'aws-s3'), Text::make('region')
+                ->textExpanderTriggers($expansions))
+            ->instructions(t('Select the region your desired bucket lives in.', category: 'aws-s3')),
+        );
+
+        // Default to an empty list:
+        $buckets = [];
+
+        try {
+            // Try and load buckets via the API:
+            $buckets = $this->listBuckets();
+        } catch (CredentialsException $e) {
+            Log::error(sprintf('Credentials were missing or invalid when trying to list the accessible buckets: %s', $e->getMessage()));
+
+            $settingsForm->add(Callout::make('bucket-list-error', t('The available credentials were not sufficient to populate a list of bucket options. If you know the bucket’s name, you can enter it manually, below.', category: 'aws-s3'))->variant('warning'));
+        }
+
+        $settingsForm->add(
+            Field::make(
+                t('Bucket', category: 'aws-s3'),
+                Combobox::make('bucket')
+                    ->options([
+                        ...Collection::make($buckets)
+                            ->map(fn(string $name): array => [
+                                'value' => $name,
+                                'label' => $name,
+                            ])
+                            ->all(),
+                        ...SelectOptions::getEnvSuggestions(),
+                    ])
+                    ->showAllOnEmpty()
+            )
+                ->instructions(t('Choose from one of the buckets accessible with the current credentials, or provide one by name or using an environment variable.', category: 'aws-s3'))
+                ->tip($environmentTip)
+        );
+
+        $settingsForm->add(
+            Field::make(t('Subfolder', category: 'aws-s3'), Text::make('subfolder')
+                ->textExpanderTriggers($expansions)
+                ->placeholder(t('path/to/subfolder', category: 'aws-s3')))
+                ->instructions(t('Your filesystem will be mounted at the root of the selected bucket, unless you provide a subpath. If you intend to use a single bucket for multiple filesystems, this is required to prevent overlap.', category: 'aws-s3'))
+        );
+
+        $settingsForm->add(
+            Field::make(t('Add the subfolder to the Base URL?', category: 'aws-s3'), Lightswitch::make('addSubfolderToRootUrl'))
+                ->instructions(t('Turn this on if you want to add the specified subfolder to the Base URL.', category: 'aws-s3'))
+        );
+
+        $settingsForm->add(
+            Field::make(t('Make Uploads Public?', category: 'aws-s3'), Lightswitch::make('makeUploadsPublic'))
+                ->instructions(t('Sets the ACL for uploaded objects. This should generally be _on_ if you want assets to be accessible by URL.', category: 'aws-s3'))
+                ->warning(t('This also applies to thumbnails that are stored on this filesystem!', category: 'aws-s3'))
+        );
+
+        if ($this->url && ! $this->makeUploadsPublic) {
+            $settingsForm->add(Callout::make('non-public-uploads-warning', t('Craft will generate URLs for assets in this filesystem, but they may not be accessible publicly.', category: 'aws-s3'))->variant('warning'));
+        }
+
+        $settingsForm->add(
+            Field::make(t('Cache Duration', category: 'aws-s3'), Text::make('expires')->monospace())
+                ->instructions(t('Used to set the `CacheControl` option when an asset is uploaded. S3 then sends the resolved value as the `Cache-Control` HTTP header when serving the asset. The value must be a valid interval expression, like `1 week` or `6 months` (a combination of a `number` and `unit` from PHP’s [relative date formatting]({url}) syntax).', ['url' => 'https://www.php.net/manual/en/datetime.formats.php#datetime.formats.relative'], 'aws-s3'))
+        );
+
+        $settingsForm->add(Separator::make('additional-settings-separator'));
+        $settingsForm->add(Heading::make('additional-services-heading', t('Additional Services', category: 'aws-s3')));
+
+        $settingsForm->add(
+            Field::make(t('Attempt to set the focal point automatically?', category: 'aws-s3'), Lightswitch::make('autoFocalPoint'))
+                ->instructions(t('Turn this on if you want to use the [AWS Rekognition]({url}) to try setting the focal point to a detected face. You can always set focal points manually.', ['url' => 'https://aws.amazon.com/rekognition/'], 'aws-s3'))
+                ->warning(t('This feature requires the `rekognition:DetectFaces` permission and may incur extra costs for each upload.', category: 'aws-s3'))
+        );
+
+        $settingsForm->add(Heading::make('cache-settings-separator', t('Cloudfront', category: 'aws-s3'))->level(3));
+
+        $settingsForm->add(
+            Field::make(t('Cloudfront Distribution ID', category: 'aws-s3'), Text::make('cfDistributionId')
+                ->textExpanderTriggers($expansions))
+                ->instructions(t('If you’re using CloudFront as a CDN for the connected bucket, enter its distribution ID so the plugin can purge assets when they’re modified.', category: 'aws-s3'))
+        );
+
+        if (! empty($this->cfDistributionId)) {
+            $settingsForm->add(
+                Field::make(t('Cloudfront Path Prefix', category: 'aws-s3'), Text::make('cfPrefix')
+                    ->textExpanderTriggers($expansions))
+                    ->instructions(t('If you’re using CloudFront as CDN for the connected bucket and have configured subfolders or custom behaviors, enter the path prefix the plugin should use when invalidating files.', category: 'aws-s3'))
+            );
+        }
+
+        return $settingsForm;
     }
 
     #[Override]
     public function getRootUrl(): ?string
     {
-        $url = parent::getRootUrl();
+        $url = Env::parse($this->url);
 
         if (! $this->addSubfolderToRootUrl) {
             return $url;
@@ -198,6 +332,20 @@ class S3 extends Filesystem
         }
 
         return $config;
+    }
+
+    public function listBuckets(): array
+    {
+        $client = new S3Client($this->getClientConfig());
+
+        $result = $client->listBuckets();
+
+        if (empty($result['Buckets'])) {
+            return [];
+        }
+
+        // Return just their names:
+        return array_column($result['Buckets'], 'Name');
     }
 
     public function detectFocalPoint(Asset $asset): array